+++ CHANGELOG.md
@@ -11,9 +11,11 @@ This release contains a LOT of breaking changes.
 
 ### Added
 
+- `static UmamiAPIClient.collect(...)` – You can collect pageviews and event without initializing a client.
 - `setDefaultUnit(...)` – Set the default unit of time when getting pageviews or events.
 - `setDefaultTZ(...)` – Set the default timezone when getting pageviews or events.
 - `setDefaultMetricType(...)` – Set the default type when getting metrics.
+- `setDefaultUserAgent(...)` – Set the default user agent when calling `collect(...)` on an instance. Does not affect the new static method.
 - `getWebsiteBy(...)` – Replaces part of the functionality of `getWebsite(...)`.
 
 ### Changed